+++ R/list-files-wd.R
@@ -7,8 +7,12 @@
 #'
 #' @param path Character vector or path. Default is "." which will set the
 #'   starting location for `relative_paths`.
-#' @param keep_non_r Logical, default is `FALSE` so keeps only records with
-#'   `relative_paths` ending in "(r|rmd|rmarkdown|qmd)$".
+#' @param pattern Regex pattern to keep only matching files. Default is
+#'   `stringr::regex("(r|rmd|rmarkdown|qmd)$", ignore_case = TRUE)` which will
+#'   keep only R, Rmarkdown and Quarto documents. To keep all files use `"."`.
+#' @param rmv_index Logical, most repos containing blogdown sites will have an
+#'   index.R file at the root. Change to `FALSE` if you don't want this file
+#'   removed.
 #'
 #' @return Dataframe with columns of `relative_paths` and `absolute_paths`.
 #' @export
@@ -23,18 +27,17 @@
 #' files_local <- list_files_wd()
 #'
 #' # Will just parse the first 2 files/gists
-#' contents <- spot_funs_files(slice(files_local, 1:2))
+#' contents <- spot_funs_files(slice(files_local, 2:3))
 #'
 #' contents %>%
 #'   unnest_results()
 #' }
 list_files_wd <- function(path = ".",
-                          keep_non_r = FALSE){
+                          pattern = stringr::regex("(r|rmd|rmarkdown|qmd)$", ignore_case = TRUE),
+                          rmv_index = TRUE) {
   contents <- tibble(relative_paths = fs::dir_ls(path = path, recurse = TRUE)) %>%
     mutate(absolute_paths = map_chr(.data$relative_paths, here::here))
 
-  if(keep_non_r){
-    return(contents)
-  } else filter(contents, str_detect_r_docs(.data$relative_paths))
+  filter(contents, str_detect_r_docs(.data$relative_paths, pattern = pattern, rmv_index = rmv_index))
 
 }