+++ NEWS.md
@@ -9,4 +9,5 @@
 # funspotr 0.0.3
 
 * Add support for quarto documents
+* No longer export `github_*` functions (instead use `list_*` funs)
 * Prepare for publishing on CRAN