+++ NEWS.md
@@ -9,6 +9,8 @@
 # funspotr 0.0.3
 
 * Add support for quarto documents
+* In `spot*files()` remove `keep_non_r` argument and add `pattern` argument
 * No longer export `github_*` functions 
   (instead use `list_*() |> spot_{pkgs|funs}_files()`)
-* Prepare for publishing on CRAN
+* Update old [blog post on gists](https://www.bryanshalloway.com/2022/02/07/identifying-r-functions-packages-in-your-github-gists/) so as to be compatible with up-to-date funspotr API
+* Publish to CRAN